requirements.txt
# ===== Web 框架 =====
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1

# ===== 多模态模型 (InternVL3.5-2B) =====
transformers>=4.40.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
Pillow>=10.0.0
einops>=0.7.0
timm>=0.9.0
sentencepiece>=0.1.99
accelerate>=0.28.0

# ===== GGUF 语言模型 (Qwen3-1.7B) =====
llama-cpp-python>=0.2.70

# ===== 向量检索 =====
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# ===== 数据库 =====
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# ===== 工具 =====
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
